games_history update
