+++ transcendence/transcendence/urls.py
@@ -22,7 +22,7 @@ from user import views
 from base import views
 from custom_auth.views import login, register, logout
 from user.views import UserCreate, UserDelete, UserGetOne, UserGetAll, UserUpdate
-from games.views import CreateGame, GameGetOne, GameGetAll, GameDelete, GameUpdate
+from games_history.views import CreateGame, GameGetOne, GameGetAll, GameDelete, GameUpdate
 from pong.views import pong
 from rest_framework_simplejwt.views import (
     TokenObtainPairView,
